gcp/cloud-functions/requirements.txt
google-cloud-storage>=2.0.0
slack-sdk>=3.0.0
